# [bounty] 2 lines attempt on NV launch overhead