Original speed on my laptop: 1.24x slower
After changes: 1.14x slower

`TORCHCUDA=1 NV=1 BEAM=3 python3 test/test_speed_v_torch.py TestSpeed.test_sub`
```
sub                             4096x 4096    1.13 ms (    14.81 GFLOPS  177.66 GB/s) in torch,    1.41 ms (    11.92 GFLOPS  143.02 GB/s) i
n tinygrad,    1.24x slower      16.78 MOPS   201.33 MB
.
----------------------------------------------------------------------
Ran 1 test in 2.603s

OK



sub                             4096x 4096    1.13 ms (    14.83 GFLOPS  178.00 GB/s) in torch,    1.29 ms (    13.02 GFLOPS  156.26 GB/s) i
n tinygrad,    1.14x slower      16.78 MOPS   201.33 MB
.
----------------------------------------------------------------------
Ran 1 test in 2.581s

OK
```

Laptop spec (3050ti)

```
❯ nvidia-smi
Sat Jul 12 18:08:38 2025       
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 570.169                Driver Version: 570.169        CUDA Version: 12.8     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  NVIDIA GeForce RTX 3050 ...    Off |   00000000:01:00.0 Off |                  N/A |
| N/A   53C    P8              6W /   60W |      15MiB /   4096MiB |      0%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+
                                                                                         
+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|    0   N/A  N/A            2160      G   ...bsx53g9-niri-25.05.1/bin/niri          2MiB |
+-----------------------------------------------------------------------------------------+
```

## More details about what is done

I first made trace of current speed with `viztracer` and got this picture
<img width="2114" height="446" alt="original_full" src="https://github.com/user-attachments/assets/0af3e02b-3c0f-4210-a471-075d0e5748ad" />


Most of the time is TinyJit and Sync, since Sync is comparable to CUDA backend, zoom in TinyJit
<img width="2114" height="446" alt="zoom_original" src="https://github.com/user-attachments/assets/b0e71192-d61a-4b22-9320-5a0e79b9647f" />

A lot of time spent in `MMIOInterface.view` and its `to_mv`
<img width="2114" height="446" alt="circles" src="https://github.com/user-attachments/assets/f3a26f32-b047-4dea-9f52-0be0654e49de" />


A bit of fiddling with `to_mv` gave a speed up

```python
- def to_mv(ptr:int, sz:int) -> memoryview: return memoryview(ctypes.cast(ptr, ctypes.POINTER(ctypes.c_uint8 * sz)).contents).cast("B")
+ def to_mv(ptr:int, sz:int) -> memoryview: return memoryview((ctypes.c_uint8 * sz).from_address(ptr)).cast("B")
```

I puzzled over why this function takes different times depending on its inputs, since it works on an existing buffer, it should take same time for every (ptr, sz)?

When size in MMIOInterface.view is not specified it takes the view till end of the buffer.
The longest time spend in the largest `view` for `fill_kernargs`, but we know its size and can specify it:

```python
argsbuf = kernargs or self.dev.kernargs_buf.offset(offset=self.dev.kernargs_offset_allocator.alloc(self.kernargs_alloc_size),
                                                   size=self.kernargs_alloc_size) # size added
```


After changes
<img width="2114" height="446" alt="zoom_new" src="https://github.com/user-attachments/assets/16413d49-c88a-47eb-af62-73d9b8a0216d" />