@nimlgen 2025-07-12
`to_mv` is used a lot and speed is important there. can you write a microbench to see how faster your impl is