@nimlgen 2025-07-12
cool, add it as a test. `class TestMemoryview(unittest.TestCase):` `test_to_mv` and you bench. and skip it for CI