@Saegl 2025-07-13
updated mv benchmark test
it doesn't really show any difference in time for different input sizes

But the first one from PR descr does, and I found another test that measures 
launch speed that shows an improvement

original
❯ NV=1 python test/test_device_speed.py
.launch 1000x 104.16 ms
launch 1000x with wait 162.56 ms


new to_mv
❯ NV=1 python test/test_device_speed.py
.launch 1000x  71.24 ms
launch 1000x with wait 101.20 ms


spec size
❯ NV=1 python test/test_device_speed.py
.launch 1000x  58.27 ms
launch 1000x with wait  90.63 ms


(new to_mv) and (spec size)
❯ NV=1 python test/test_device_speed.py
.launch 1000x  47.77 ms
launch 1000x with wait  76.54 ms